+++ src/TrustedHostsNetworkResolver.php
@@ -12,10 +12,9 @@ use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use RuntimeException;
 use Yiisoft\Http\HeaderValueHelper;
+use Yiisoft\NetworkUtilities\IpRanges;
 use Yiisoft\ProxyMiddleware\Exception\InvalidConnectionChainItemException;
 use Yiisoft\ProxyMiddleware\Exception\RfcProxyParseException;
-use Yiisoft\Validator\Rule\Ip;
-use Yiisoft\Validator\ValidatorInterface;
 
 use function count;
 use function in_array;
@@ -130,7 +129,6 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
 
     /**
      * @psalm-var list<non-empty-string>
-     * @psalm-suppress PropertyNotSetInConstructor
      */
     private array $trustedIps = [];
     /**
@@ -146,7 +144,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
      */
     private ?string $connectionChainItemsAttribute = null;
 
-    public function __construct(private ValidatorInterface $validator)
+    public function __construct()
     {
     }
 
@@ -165,7 +163,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
         foreach ($trustedIps as $ip) {
             $this->assertIsNonEmptyString($ip, 'Trusted IP');
 
-            if (!$this->isIp($ip)) {
+            if (!IpValidator::isIp($ip)) {
                 throw new InvalidArgumentException("\"$ip\" is not a valid IP.");
             }
 
@@ -672,7 +670,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
                 if (isset($matches['ipv6']) && !empty($matches['ipv6'])) {
                     $ip = $matches['ipv6'];
 
-                    if (!$this->isIpv6($ip)) {
+                    if (!IpValidator::isIpV6($ip)) {
                         $message = "Enclosing in square brackets assumes presence of valid IPv6, \"$ip\" given.";
 
                         throw new RfcProxyParseException($message);
@@ -722,7 +720,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
         bool $validateIp = true,
         bool $validateProtocol = true,
     ): array {
-        if ($ip !== null && $validateIp && !$this->isIp($ip)) {
+        if ($ip !== null && $validateIp && !IpValidator::isIp($ip)) {
             throw new InvalidConnectionChainItemException("\"$ip\" is not a valid IP.");
         }
 
@@ -847,31 +845,9 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
             }
         }
 
-        if (!$this->isIp($ipData[0])) {
+        if (!IpValidator::isIp($ipData[0])) {
             throw new RuntimeException('IP returned from reverse-obfuscated IP data is not valid.');
         }
-    }
-
-    /**
-     * @psalm-assert non-empty-string $value
-     */
-    private function isIp(string $value): bool
-    {
-        return $this
-            ->validator
-            ->validate($value, [new Ip()])
-            ->isValid();
-    }
-
-    /**
-     * @psalm-assert non-empty-string $value
-     */
-    private function isIpv6(string $value): bool
-    {
-        return $this
-            ->validator
-            ->validate($value, [new Ip(allowIpv4: false)])
-            ->isValid();
     }
 
     /**
@@ -879,7 +855,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
      */
     private function isPrivateIp(string $value): bool
     {
-        return (new Ip(ranges: ['private']))->isAllowed($value);
+        return IpValidator::inRanges($value, [IpRanges::PRIVATE]);
     }
 
     /**
@@ -887,7 +863,7 @@ class TrustedHostsNetworkResolver implements MiddlewareInterface
      */
     private function isTrustedIp(string $value): bool
     {
-        return !empty($this->trustedIps) && (new Ip(ranges: $this->trustedIps))->isAllowed($value);
+        return !empty($this->trustedIps) && IpValidator::inRanges($value, $this->trustedIps);
     }
 
     /**